问题风险修正

+++ src/main/java/ext/st/pmgt/issue/processors/DeleteRiskMeasuresProcessor.java
@@ -22,11 +22,20 @@ import java.util.List;
 public class DeleteRiskMeasuresProcessor extends DefaultObjectFormProcessor {
     @Override
     public ResponseWrapper<?> doOperation(ComponentParams params, List list) throws PIException {
-        List<STProjectMeasures> measures=(List)params.getNfCommandBean().getSelectedObjects();
-        STProjectRisk risk = (STProjectRisk)params.getNfCommandBean().getSourceObject();
+        List<STProjectMeasures> measures=(List)params.getNfCommandBean().getSelectedObjects();//批量按钮删除
+        Persistable sourceObject = params.getNfCommandBean().getSourceObject();
         for (STProjectMeasures projectMeasures:measures){
             PersistenceHelper.service.delete(projectMeasures);
             ContentHolder contentHolder=projectMeasures;
+            List<ContentHolder> contentHolderList=new ArrayList<>();
+            contentHolderList.add(contentHolder);
+            PIProjectChangeHelper.service.deleteAttachmentData(contentHolderList);
+        }
+
+        if (sourceObject instanceof STProjectMeasures){  //编辑删除
+            STProjectMeasures stProjectMeasures=(STProjectMeasures)sourceObject;
+            PersistenceHelper.service.delete(stProjectMeasures);
+            ContentHolder contentHolder=stProjectMeasures;
             List<ContentHolder> contentHolderList=new ArrayList<>();
             contentHolderList.add(contentHolder);
             PIProjectChangeHelper.service.deleteAttachmentData(contentHolderList);
